requirements.txt
anyio==4.4.0
astor==0.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
camelot-py==0.11.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
distro==1.9.0
easyocr==1.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.6.1
ghostscript==0.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.5.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.0
paddlepaddle==2.6.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.1
pdftopng==0.2.3
pillow==10.4.0
protobuf==5.27.2
pyclipper==1.3.0.post5
pycparser==2.22
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12.1
tabula-py==2.9.3
tabulate==0.9.0
termcolor==2.4.0
tifffile==2024.7.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
